bots/openai-bot/requirements.txt
bubbletea-chat[llm]
openai
httpx
python-dotenv
